requirements.txt
python-terrier
fastapi
pandas
uvicorn
pydantic
nltk
beautifulsoup4
python-dotenv
autocomplete
psycopg2
sqlalchemy